Library

MORFEInvariantManifold.BE2n — Constant

Number of nodes of a linear line element (2)

MORFEInvariantManifold.BE3n — Constant

Number of nodes of a quadratic line element (3)

MORFEInvariantManifold.H20n — Constant

Number of nodes of a quadratic serendipity hexahedral element (20)

MORFEInvariantManifold.H27n — Constant

Number of nodes of a quadratic Lagrange hexahedral element (27)

MORFEInvariantManifold.HE8n — Constant

Number of nodes of a linear hexahedral element (8)

MORFEInvariantManifold.P15n — Constant

Number of nodes of a quadratic 15-nodes wedge element (15)

MORFEInvariantManifold.P18n — Constant

Number of nodes of a quadratic 18-nodes wedge element (18)

MORFEInvariantManifold.PE6n — Constant

Number of nodes of a linear wedge element (6)

MORFEInvariantManifold.QU4n — Constant

Number of nodes of a linear square element (4)

MORFEInvariantManifold.QU8n — Constant

Number of nodes of a quadratic serendipity square element (8)

MORFEInvariantManifold.QU9n — Constant

Number of nodes of a quadratic Lagrange square element (9)

MORFEInvariantManifold.Snse — Constant

Number of supported boundary element

MORFEInvariantManifold.T10n — Constant

Number of nodes of a quadratic tetrahedral element (10)

MORFEInvariantManifold.TE4n — Constant

Number of nodes of a linear tetrahedral element (4)

MORFEInvariantManifold.TR3n — Constant

Number of nodes of a linear triangular element (3)

MORFEInvariantManifold.TR6n — Constant

Number of nodes of a quadratic triangular element (6)

MORFEInvariantManifold.VE1n — Constant

Number of nodes of a vertex (1)

MORFEInvariantManifold.Vnse — Constant

Number of supported volume element

MORFEInvariantManifold.dim — Constant

Dimension of the treated problems (3D)

MORFEInvariantManifold.ncv_max — Constant

Maximum number of elements connected one to the other

MORFEInvariantManifold.nne_max — Constant

Maximum number nodes for an element

MORFEInvariantManifold.ϵ_tol — Constant

resonances tolerance

MORFEInvariantManifold.Boundary — Type

mutable struct Boundary

Sen : number of supported elements

Set : type of supported elements

Senn : number of nodes of the supported elements

ne : number of elements of each type in the domain

e2n : surface element connectivity

bcdofs : which dofs are constrained

bcvals : value associated to the boundary condition

MORFEInvariantManifold.Domain — Type

mutable struct Domain

Sen : number of supported elements

Set : type of supported elements

Senn : number of nodes of the supported elements

ne : number of elements of each type in the domain

e2n : volume element connectivity

mat : material associated to the domain

MORFEInvariantManifold.Field — Type

mutable struct Field

neq : number of equations

nen : total number of entries

cpn : nodal uknowns per node

dof : dofs ordering

val : field values

MORFEInvariantManifold.Field — Method

Field(mesh::Grid, cpn::Int64)

Field constructor

mesh : Grid

cpn : number of nodal unknowns per node

MORFEInvariantManifold.Grid — Type

mutable struct grid

netype : number of supperted types of elements

etype : array that stores the number of nodes for each element

nn : number of nodes

nΩ : number of domains

nΓ : number of boundaries

n2c : nodes coordinates

Ω : domains

Γ : boundaries

tcn : total cells number

grid() : empty initialization method

MORFEInvariantManifold.NodalConnectivity — Type

struct NodalConnectivity

nval : number of nodes + 1

nreg : total size of the nodal connectivity

register : bookkeeping nodes to nodes connectivity in compact sparse format

values : stored node to node connectivity

maxr : FILL

MORFEInvariantManifold.NodalConnectivity — Method

NodalConnectivity(mesh::Grid)

NodalConnectivity main constructor.

mesh : Grid

MORFEInvariantManifold.Parametrisation — Type

mutable struct Parametrisation

nc : number of monomial combinations

W : mappings [Υ, Ψ]

f : reduced dynamics

nlr : nonlinear right hand side

comb : combinations the system is solved for

conj : conjugacy relations

cmap : mapping index combinations

MORFEInvariantManifold.material — Type

struct material

ρ : density of the material

Dᵢⱼₖₗ : fourth order elasticity tensor in Voigt notation

MORFEInvariantManifold.MORFE_mech_autonomous — Function

MORFEmechautonomous(meshfile,domainslist,materials,boundarieslist,constraineddof,bcvals,α,β,Φₗᵢₛₜ,style,maxorder,neig=0,nls=0)

Parametrisation of the autonomous system:

\[\mathbf{M}\ddot{\mathbf{U}}+\mathbf{C}\dot{\mathbf{U}}+\mathbf{F(\mathbf{U})}=0\]

mesh_file : name of the mesh file with extension

domains_list : tags of domains included in the analysis

materials : materials associated to each domain

boundaries_list : list of all boundaries included in the analysis

constrained_dof : list of constrained degrees of freedom

bc_vals : value of the boundary condition

α : mass-proportional damping coefficient

β : stiffness-proportional damping coefficient

Φₗᵢₛₜ : list of the master modes included in the analysis

style : parametrisation style

max_order : maximum order of the asymptotic development

neig : number of eigenvalues. If not used, then neig = maximum(Φₗᵢₛₜ)

nls : nonlinear static analysis. 0 = no, 1 = yes.

MORFEInvariantManifold.MORFE_mech_nonautonomous — Function

MORFEmechautonomous(meshfile,domainslist,materials,boundarieslist,constraineddof,bcvals,Ωlist,κmodes,κlist,κphase,α,β,Φₗᵢₛₜ,style,maxordera,maxorder_na,neig=0,nls=0)

Parametrisation of the autonomous system:

\[\mathbf{M}\ddot{\mathbf{U}}+\mathbf{C}\dot{\mathbf{U}}+\mathbf{F(\mathbf{U})}=\mathbf{F}(t)\]

mesh_file : name of the mesh file with extension

domains_list : tags of domains included in the analysis

materials : materials associated to each domain

boundaries_list : list of all boundaries included in the analysis

constrained_dof : list of constrained degrees of freedom

bc_vals : value of the boundary condition

Ω_list : list of forcing frequencies

κ_modes : load multiplier associated to each forcing term

κ_list : shape of each forcing

κ_phase : identification of the forcing phase. "c" = cosine, "s" = sine

α : mass-proportional damping coefficient

β : stiffness-proportional damping coefficient

Φₗᵢₛₜ : list of the master modes included in the analysis

style : parametrisation style

max_order : maximum order of the asymptotic development

neig : number of eigenvalues. If not used, then neig = maximum(Φₗᵢₛₜ)

nls : nonlinear static analysis. 0 = no, 1 = yes.

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:H20}})

It fills N with the shape functions of a PE6 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:H20}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:H27}})

It fills N with the shape functions of a PE6 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:H27}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:HE8}})

It fills N with the shape functions of a H8 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:HE8}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:P15}})

It fills N with the shape functions of a PE6 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:P15}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:P18}})

It fills N with the shape functions of a PE6 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:P18}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:PE6}})

It fills N with the shape functions of a PE6 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:PE6}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:T10}})

It fills N with the shape functions of a T10 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:T10}} : element type specifier

MORFEInvariantManifold.N! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:TE4}})

It fills N with the shape functions of a T4 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:TE4}} : element type specifier

MORFEInvariantManifold.N27! — Method

N!(N::Array{Float64},gp::Tuple,::Type{Val{:H27}})

It fills N with the shape functions of a PE6 element evaluated at gauss point gp.

N : shape functions Array

gp : gauss point coordinates

:Type{Val{:H27}} : element type specifier

MORFEInvariantManifold.add_material — Method

add_material( name::String, ρ::Float64, E::Float64, ν::Float64 )

It adds an isotropic material to the internal database of the package.

name = name of the material

ρ = density of the material

E = Young's modulus

ν = Poisson's ratio

return nothing

MORFEInvariantManifold.add_material — Method

add_material( name::String, ρ::Float64, Dᵢⱼₖₗ::Matrix{Float64} )

It adds an anisotropic material to the internal database of the package.

name = name of the material

ρ = density of the material

Dᵢⱼₖₗ = fourth order elasticity tensor

return nothing

MORFEInvariantManifold.append_rdyn_frequency! — Method

appendrdynfrequency!(rdyn,ndofs,ith_Ω)

It appends auxiliary variables required to recast the non-autonomous reduced dynamics as autonomous

rdyn : reduced dynamics

ndofs : number of degrees of freedom

ith_Ω : integer represeting which excitation frequency you are exciting

MORFEInvariantManifold.assembly! — Method

assembly!(K::SparseMatrixCSC, val::Float64, ir::Int64, jc::Int64)

Assembly value in sparse matrix using binary search

K : global matrix

val : value to assemble

ir : row entry

jc : column entry

MORFEInvariantManifold.assembly_G_nl! — Function

assemblyGnl!(Cp, entry,Ψ₁, Ψ₂,mesh, U, mult = 1.0)

It assemblies quadratic nonlinearities operator

$ G(Ψ₁,Ψ₂) = \frac{1}{2} ∫_{Ω} γ(Ψ₁,Ψ₂):\mathcal{A}:ε(w) + γ(Ψ₁,w):\mathcal{A}:ε(Ψ₂) + γ(w,Ψ₂):\mathcal{A}:ε(Ψ₁) dΩ $

Cp : parametrisation data structure

entry : entrance of the reference array

Ψ₁ : mapping

Ψ₂ : mapping

mesh : mesh data structure

U : displacement field

mult : integral multiplier

MORFEInvariantManifold.assembly_H_nl! — Function

assemblyHnl!(Cp, entry,Ψ₁, Ψ₂, Ψ₃,mesh, U, mult = 1.0)

It assemblies cubic nonlinearities operator

$ G(Ψ₁,Ψ₂,Ψ₃) = \frac{1}{6} ∫_{Ω} γ(Ψ₁,Ψ₂):\mathcal{A}:γ(Ψ₃,w) + γ(Ψ₁,Ψ₃):\mathcal{A}:γ(Ψ₂,w) + γ(Ψ₃,Ψ₂):\mathcal{A}:γ(Ψ₁,w) dΩ $

Cp : parametrisation data structure

entry : entrance of the reference array

Ψ₁ : mapping

Ψ₂ : mapping

Ψ₃ : mapping

mesh : mesh data structure

U : displacement field

mult : integral multiplier

MORFEInvariantManifold.assembly_MCK! — Method

assembly_MCK!(mesh::Grid, U::Field, M::SparseMatrixCSC, C::SparseMatrixCSC, K::SparseMatrixCSC, α::Float64, β::Float64)

It assemblies mass, damping, and stiffness matrices

mesh : Grid

U : displacement field

M : mass matrix

C : damping matrix

K : stiffness matrix

α : Rayleigh damping mass proportional coefficient

β : Rayleigh damping stiffness proportional coefficient

MORFEInvariantManifold.assembly_cubic_nl! — Method

assemblycubicnl!(Cp,entry,ic,ndofs,p,mesh,U)

It assembles cubic nonlinearities vector for a given monomial I of order p

Cp : parametrisation data structure

entry : index entry

ic : index combination of the monomial

ndofs : number or degrees of freedom

p : order of the asymptotic development

mesh : grid data structure

U : displacement field

MORFEInvariantManifold.assembly_cubic_nl_veps! — Method

assemblycubicnl_veps!(Cp,Cp⁺,entry,Cic,ndofs,p,mesh,U)

It computes cubic nonlinearity vectors for all monomials of order p of the non-autonomous asymptotic development

Cp : autonomous parametrisation data structure

Cp⁺ : non-auonomous parametrisation data structure

entry : array entry of a given monomial

ic : index combination associated to a monomial

ndofs : number of degrees of freedom

p : order of the asymptotic development

mesh : grid data structure

U : displacement field

MORFEInvariantManifold.assembly_forcing_vector! — Method

assemblyforcingvector!(Cp⁺,mesh,U,M,ϕ,κmodes,κlist,κ_phase)

It assemblies the non-autonomous forcing vector

$ \mathbf{F} = ∑_{i=1}^{N} κᵢMϕᵢ $

Cp⁺ : no-autonomous parametrisation data structure

mesh : mesh data structure

U : displacement field

M : mass matrix

ϕ : eigenmodes

κ_modes : list of forced modes

κ_list : forcing amplitude κᵢ

κ_phase : phase of the applied forcing

MORFEInvariantManifold.assembly_mat_sym! — Method

assemblymatsym!(K::SparseMatrixCSC,Kₑ::Matrix{Float64},dofs::Array{Int64},rl::Int64)

It assembles elemental matrices into sparse matrices. Is assumes CSC format with lower triangular.

K : global matrix

Kₑ : elemental matrix

dofs : degrees of freedom

rl : row length

MORFEInvariantManifold.assembly_quadratic_nl! — Method

assemblyquadraticnl!(Cp,entry,ic,ndofs,p,mesh,U,nls)

It assembles quadratic nonlinearities vector for a given monomial I of order p

Cp : parametrisation data structure

entry : index entry

ic : index combination of the monomial

ndofs : number or degrees of freedom

p : order of the asymptotic development

mesh : grid data structure

U : displacement field

nls : tag for nonlinear static analysis. 0 = no, 1 = yes

MORFEInvariantManifold.assembly_quadratic_nl_veps! — Method

assemblycubicnl_veps!(Cp,Cp⁺,entry,Cic,ndofs,p,mesh,U)

It computes quadratic nonlinearity vectors for all monomials of order p of the non-autonomous asymptotic development

Cp : autonomous parametrisation data structure

Cp⁺ : non-auonomous parametrisation data structure

entry : array entry of a given monomial

ic : index combination associated to a monomial

ndofs : number of degrees of freedom

p : order of the asymptotic development

mesh : grid data structure

U : displacement field

nls : nonlinear static solution. 0 = no, 1 = yes

MORFEInvariantManifold.assembly_sys_mat! — Method

assemblysysmat!(Cp,sysmat,M,C,K,σ,resonantmodes,ndofs)

it assembles the linear system associated to a homological equation

Cp : parametrisation structure

sys_mat : matrix associated to the homological equation

M : mass matrix

C : damping matrix

K : stiffness matrix

σ : summation of eigenvalues λ

resonant_modes : book-keeping of the set R of resonant modes

ndofs : number of degrees of freedom

MORFEInvariantManifold.assembly_sys_rhs! — Method

assemblysysrhs!(Cp,sysrhs,sysres,M,C,comb,p,ndofs,σ,resonant_modes)

It assembles the right hand side of a given homological equation of the ε⁰-developmet

Cp : parametrisation structure

sys_rhs : right hand side of the homological equations

sys_res : residual of the homological equations

M : mass matrix

C : damping matrix

comb : reference monomial index

p : order of the asymptotic development

ndofs : number of degrees of freedom

σ : eingevalues λ summation

resonant_modes : book-keeping of the set R of resonant modes

MORFEInvariantManifold.assembly_sys_rhs_veps! — Method

assemblysysrhs!(Cp,Cp⁺,sysrhs,sysres,M,C,comb,p,ndofs,σ,resonant_modes)

It assembles the right hand side of a given homological equation of the ε⁰-developmet

Cp : autonomous parametrisation data structure

Cp⁺ : non-auonomous parametrisation data structure

sys_rhs : right hand side of the homological equations

sys_res : residual of the homological equations

M : mass matrix

C : damping matrix

comb : reference monomial index

p : order of the asymptotic development

ndofs : number of degrees of freedom

σ : eingevalues λ summation

resonant_modes : book-keeping of the set R of resonant modes

MORFEInvariantManifold.assembly_μ_ν_resid! — Method

assemblyμν_resid!(Cp,entry,ic,ndofs,p,neq)

It assembles μ and ν for a given monomial I of order p for the ε⁰ development.

Cp : parametrisation data structure

entry : entry associated to a monomial combination

ic : monomial index combination

ndofs : number of degrees of freedom

p : order of the asymptotic development

neq : number of equations

MORFEInvariantManifold.assembly_μ_ν_resid_veps! — Method

recursiveassemblyveps!(Cp,Cp⁺,ndofs,p,mesh,U,Cic,cc,nls=0)

It assembles right hand sides of all homological equations for a given order p of the expansion.

Cp : autonomous parametrisation data structure

Cp⁺ : non-auonomous parametrisation data structure

entry : array entry of a given monomial

ic : index combination associated to a monomial

ndofs : number of degrees of freedom

p : order of the asymptotic development

neq : number of equations

MORFEInvariantManifold.check_resonances! — Method

checkresonances!(Cp,σ,ndofs,style,resonantmodes)

It checks which mode is resonant. It complies with the chosen style.

Cp : parametrisation data structure

σ : summation of eigenvalues λ

ndofs : number of degrees of freedom

style : parametrisation style

resonant_modes : book-keeping of the set R of resonant modes

MORFEInvariantManifold.compute_ncomb — Method

compute_ncomb(param,order,ndofs,neq)

It computes the monomials combinations that need to be computed

p : order of the expansion

ndofs : number of degrees of freedom of the reduced model

MORFEInvariantManifold.delete_material — Method

delete_materials()

It delets a material in the database.

name = name of the material to delete

return nothing

MORFEInvariantManifold.dofs! — Method

dofs!(ϕ::Field, nn::Int64, nodes::SubArray{Int64, 1, Vector{Int64}, Tuple{UnitRange{Int64}}, dofs::Vector{nt64})

It copies the dofs associated to nodes list nodes into dofs. This version accepts SubArrays.

ϕ : field of interest

nn : number of queries

nodes : list of nodes

dofs : number of degrees of freedom

MORFEInvariantManifold.dofs! — Method

dofs!(ϕ::Field, nn::Int64, nodes::Vector{Int64}, dofs::Vector{nt64})

It copies the dofs associated to nodes list nodes into dofs.

ϕ : field of interest

nn : number of queries

nodes : list of nodes

dofs : number of degrees of freedom

MORFEInvariantManifold.dofs_vals! — Method

dofs_vals!(ϕ::Field, nn::Int64, nodes::Vector{Int64}, dofs::Vector{Int64}, vals::Vector{Float64})

It copies the dofs and values associated to nodes list nodes into dofs and vals.

ϕ : field of interest

nn : number of queries

nodes : list of nodes

dofs : number of degrees of freedom

vals : values of the field

MORFEInvariantManifold.export_C — Method

export_C(odir,coeff,ndofs)

it saves the monomials exponents.

odir : output directory

coeff : exponents of the monomial

ndofs : number of degrees of freedom of the reduced model

MORFEInvariantManifold.export_M — Method

export_M(Wr,comb,odir,M,Φ,neig)

it saves the mass-wieighted projections of the mappings onto the computed modal basis.

Wr : realified mapping of order p

comb : saved index combination

odir : output directory

M : mass-matrix

Φ : eigenmodes

neig : number of computed eigenmodes

MORFEInvariantManifold.export_W — Method

exportW(Ψ,Υ,coeff,U,odirW)

it saves the realified mappings

Ψ : realified displacement mapping

Υ : realified velocity mapping

odir : output directory

MORFEInvariantManifold.export_data! — Method

exportdata!(Cp,Wr,fr,p,ndofs,U,M,Φ,neig,odirC,odirW,odirf,odir_M)

It saves the output of a given order of the parametrisation procedure. Only realified quantities are saved.

Cp : order p of a given parametrisation

Wr : realified mapping of order p

fr : realified reduced dynamics of order p

p : order of the asymptotic expansion

ndofs : number of degrees of freedom

U : displacement field

M : mass matrix

Φ : eigenmodes list

neig : number of computed eigenvalues

odir_C : path to the folder that stores monomials exponents

odir_W : path to the folder that stores the realified mappings

odir_f : path to the folder that stores the realified reduced dynamics

odir_M : path to the folder that stores the mass-wiedghted projected mappings onto the computed modes

MORFEInvariantManifold.export_data_veps! — Method

exportdataveps!(Cp,Wr,fr,p,ndofs,U,M,Φ,neig,odirCc, odirWc, odirfc, odirMc, odirCs, odirWs, odirfs, odirMs)

It saves the output of a given order of the non-autonomous parametrisation procedure. Only realified quantities are saved.

Cp : order p of a given parametrisation

Wr : realified mapping of order p

fr : realified reduced dynamics of order p

p : order of the asymptotic expansion

ndofs : number of degrees of freedom

U : displacement field

M : mass matrix

Φ : eigenmodes list

neig : number of computed eigenvalues

odirCc : path to the folder that stores monomials exponents, cosine harmonic

odirWc : path to the folder that stores the realified mappings, cosine harmonic

odirfc : path to the folder that stores the realified reduced dynamics, cosine harmonic

odirMc : path to the folder that stores the mass-wiedghted projected mappings onto the computed modes, cosine harmonic

odirCs : path to the folder that stores monomials exponents, sine harmonic

odirWs : path to the folder that stores the realified mappings, sine harmonic

odirfs : path to the folder that stores the realified reduced dynamics, sine harmonic

odirMs : path to the folder that stores the mass-wiedghted projected mappings onto the computed modes, sine harmonic

MORFEInvariantManifold.export_eig — Method

exporteig(meshfile::String, U::fieldtype, λ::Vector{ComplexF64},ϕ::Matrix{ComplexF64},outdir::String)

It exports the eigenfunctions in vtk format and eigenfrequencies in .txt file

mesh_file : name of the mesh file

U : dummy field variable

λ : eigenvalues

ϕ : eigenfunctions

out_dir : output directory

MORFEInvariantManifold.export_f — Method

export_f(fr,ndofs,odir)

it saves the realified reduced dynamics.

fr : realified reduced dynamics vector

ndofs : number of degrees of freedom of the reduced model

odir : output directory

MORFEInvariantManifold.fill_comb! — Method

fillcomb!(comb::Matrix{Int64},cmap::Array{Int64},ndofs::Int64,p::Int64,combs::Array{Int64},ithc::Int64,counter::Int64)

It fills the comb matrix with the associated exponentials

comb : matrix that stores the powers of the monomials

cmap : total combinations map

ndofs : number of degrees of freedom

p : order of the asymptotic expansion

combs : combinations

ith_c : ith-combination

counter : recursion depth

MORFEInvariantManifold.fill_mapping! — Method

fill_mapping!(Ψ,Υ,Wr,e,U)

it copies the mappings but taking into account the global nodes ordering

Ψ : realified displacement mapping

Υ : realified velocity mapping

Wr : realified mapping

e : combination index

U : displacement field

MORFEInvariantManifold.fill_mapping_c! — Method

fill_mapping!(Ψ,Υ,Wr,e,U)

it copies the mappings but taking into account the global nodes ordering. Cosine variant.

Ψ : realified displacement mapping

Υ : realified velocity mapping

Wr : realified mapping

e : combination index

U : displacement field

MORFEInvariantManifold.fill_mapping_s! — Method

fill_mapping!(Ψ,Υ,Wr,e,U)

it copies the mappings but taking into account the global nodes ordering. Since variant.

Ψ : realified displacement mapping

Υ : realified velocity mapping

Wr : realified mapping

e : combination index

U : displacement field

MORFEInvariantManifold.fill_rdyn! — Method

init_rdyn(ndofs)

It initializes the real-valued reduced dynamics in a format compatible with MATCONT

ndofs : number of degrees of freedom

MORFEInvariantManifold.fill_rdyn_c! — Method

fillrdync!(rdyn,fr,i,ndofs)

Harmonics realification step. Cosine variant.

rdyn : realified reduced dynamics

fr : reduced dynamics

i : index combination

ndofs : number of degrees of freedom

MORFEInvariantManifold.fill_rdyn_s! — Method

fillrdyns!(rdyn,fr,i,ndofs)

Harmonics realification step. Sine variant.

rdyn : realified reduced dynamics

fr : reduced dynamics

i : index combination

ndofs : number of degrees of freedom

MORFEInvariantManifold.fill_rdyn_veps! — Method

fillrdynveps!(rdyn,ndofs,fr,Cp,p,ith_Ω)

rdyn : reduced dynamics

ndofs : number of degrees of freedom

fr : realified reduced dynamics

Cp : parametrisation of order p

p : order of the asymptotic development

ith_Ω : integer represeting which excitation frequency you are exciting

MORFEInvariantManifold.find_index_global — Method

findindexglobal(combs,ndofs,p)

It computes the position in the cmap array of a given combination

combs : total combinations

ndofs : number of degrees of freedom

p : order of the expansion

MORFEInvariantManifold.find_index_map — Method

findindexmap(j::Int64,k::Int64,s::Int64,p::Int64,ndofs::Int64,index_comb::Array{Int64})

It finds the position of a map for a certain index_comb during time-derivatives residual calculation.

j : counter

k : counter

s : counter

p : expansion order

ndofs : number of degrees of freedom

index_comb : reference monomials exponentials

MORFEInvariantManifold.find_index_rdyn — Method

findindexrdyn(j::Int64,k::Int64,ndofs::Int64,index_comb::Array{Int64})

It finds the position of a reduced dynamics vector for a certain index_comb during time-derivatives residual calculation.

j : counter

k : counter

ndofs : number of degrees of freedom

index_comb : reference monomials exponentials

MORFEInvariantManifold.find_position_g — Method

findpositiong(j::Int64,p::Int64,ndofs::Int64,index_comb::Array{Int64})

It finds the maps position to compute their contribution to the quadratic nonlinearities.

j : counter

p : order

ndofs : number of degrees of freedom

index_comb : reference monomials exponentials

MORFEInvariantManifold.find_position_h — Method

findpositiong(j::Int64,k::Int64,p::Int64,ndofs::Int64,index_comb::Array{Int64})

It finds the maps position to compute their contribution to the cubic nonlinearities.

j : counter

k : counter

p : order

ndofs : number of degrees of freedom

index_comb : reference monomials exponentials

MORFEInvariantManifold.generate_conj! — Method

generate_conj!(conj,comb,cmap,ndofs,p,combs,nc)

It maps entries to their conjugate

conj : conjugate

comb : matrix that stores the powers of the monomials

cmap : total combinations map

ndofs : number of degrees of freedom

p : order of the asymptotic expansion

combs : combinations

nc : number of combinations

MORFEInvariantManifold.generate_pmap! — Method

generatepmap!(comb::Matrix{Int64},cmap::Array{Int64},ndofs::Int64,p::Int64,combs::Array{Int64},sortedcombs::Array{Int64},counter::Int64)

It computes combinations maps to account for indexes permutations.

comb : matrix that stores the powers of the monomials

cmap : total combinations map

ndofs : number of degrees of freedom

p : order of the asymptotic expansion

combs : combinations

sorted_combs : sorted combinations

counter : recursion depth

MORFEInvariantManifold.get_coor! — Method

get_coor!(mesh::Grid, conn::SubArray{Int64, 1, Vector{Int64}, Tuple{UnitRange{Int64}}, true}, X::Vector{Float64}, nn::Int64)

it retrieves element nodes coordinates.

mesh : Grid

conn : connectivity of the element

X : coordinates of the nodes in {x₁, y₁, z₁, ..., xₙ, yₙ, zₙ} format

nn : number of nodes

MORFEInvariantManifold.get_coor! — Method

get_coor!(mesh::Grid, conn::Vector{Int64}, X::Vector{Float64}, nn::Int64)

it retrieves element nodes coordinates.

mesh : Grid

conn : connectivity of the element

X : coordinates of the nodes in {x₁, y₁, z₁, ..., xₙ, yₙ, zₙ} format

nn : number of nodes

MORFEInvariantManifold.get_materials_path — Method

getmaterialspath()

It returns the path associated to the material database return pth::String

MORFEInvariantManifold.get_node_connectivity — Method

NodalConnectivity(n2n::NodalConnectivity, node::Int64)

It returns a View containing which nodes are connector to node

n2n : NodalConnectivity structure

node : node of interest

MORFEInvariantManifold.identity_tangency! — Method

identity_tangency!(Cp::Parametrisation,nm::Int64,neq::Int64,Φₗᵢₛₜ::Array{Int64},ϕ::Matrix{ComplexF64},ω₀::Array{Float64},ζ₀::Array{Float64})

It fills the first order parametrization using eigenmodes and eigenvalues of the mechanical problem.

Cp : Parametrisation data structure

nm : number of master modes

neq : total number of equations

Φₗᵢₛₜ : master modes list

ϕ : modes

ω₀ : eigenfrequencies

ζ₀ : damping coefficients

U : displacement field

MORFEInvariantManifold.import_mesh! — Method

importmesh!( mesh::Grid, meshfile::String, Ωlist::Vector{Vector{Int64}}, Γlist::Vector{Vector{Int64}}, mat::Vector{material}, bcdof::Vector{Vector{Int64}}, bcvals::Vector{Vector{Float64}}, ::Type{Val{:MED}})

It reads a MED mesh format

mesh_file : name of the mesh file

Ω_list : list of domains in which the structures is organised

Γ_list : list of boundaries

mat : list of materials associated to each domain

bc_dof : list of constrained degrees of freedom of each boundary

bc_vals : boundary condition values

:MED : specializes function to read MED mesh format

MORFEInvariantManifold.increment_field! — Method

increment_field!(ϕ::Field,U::Vector{Float64})

it increments field values. ϕ.val += U

ϕ : field to update

U : solution vector

MORFEInvariantManifold.init_rdyn — Method

init_rdyn(ndofs)

It initializes the real-valued reduced dynamics in a format compatible with MATCONT

ndofs : number of degrees of freedom

MORFEInvariantManifold.init_realification_matrix — Method

initrealificationmatrix(ndofs::Int64)

It initializes the matrix used to realify the system

MORFEInvariantManifold.init_symCSC — Method

init_symCSC(mesh::Grid, ϕ::Field, n2n::NodalConnectivity, naux::Int64, tp::String)

It initializes a symmetric CSC matrix from a field. It eventually add auxiliary equations.

mesh : Grid

ϕ : field of interest

nn : number of queries

n2n : node-to-node connectivity

naux : number of auxiliary equations

MORFEInvariantManifold.initialize_parametrisation! — Function

initialize_parametrisation!(param,order,ndofs,neq)

It initializes a generic order of the asymptotic expansion

param : parametrisation

order : order

ndofs : number of degrees of freedom of the reduced model

neq : number of equations

MORFEInvariantManifold.integrate_MK! — Method

integrate_MK!(Mₑ::Matrix{Float64},Kₑ::Matrix{Float64},X::Array{Float64},

ρ::Float64,Dᵢⱼₖₗ::Matrix{Float64}, N::Array{Float64},∂N∂a::Matrix{Float64},∂N∂x::Matrix{Float64}, sym∇::Matrix{Float64},Jac::Matrix{Float64}, Jac⁻¹::Matrix{Float64},nn,etype::Symbol) It integrates elemental and stiffness matrices

Mₑ : elemental mass matrix

Kₑ : elemental stiffness matrix

X : nodal coordinates

ρ : density

Dᵢⱼₖₗ : elasticity tensor

N : shape functions

∂N∂a : shape functions master derivatives

∂N∂x : shape functions sparial derivatives

sym∇ : 0.5*(∇(⋅)+ ∇ᵀ(⋅))

Jac : ∂x∂a

Jac⁻¹: ∂a∂x

nn : nodes number

etype : element type

MORFEInvariantManifold.list_materials — Method

list_materials()

It lists all materials in the database. return nothing

MORFEInvariantManifold.load_material — Method

load_material( name::String )

It loads the data structure from the database

name : name of the material to load

MORFEInvariantManifold.make_output_dir — Method

makeoutputdir(mesh_file::String)

It creates the folder that stores the output of the analysis

mesh_file : name of the mesh file

MORFEInvariantManifold.mass_normalization! — Method

mass_normalization!(ϕ,M,neig)

Modes are normalised such that ϕᵢMϕⱼ=δᵢⱼ

ϕ : eigenmodes

M : mass matrix

neig : number of computed eigenvalues

MORFEInvariantManifold.metric! — Method

metric!(X::Array{Float64},∂N∂a::Matrix{Float64},

      ∂N∂x::Matrix{Float64},nn::Int64,
      Jac::Matrix{Float64},Jac⁻¹::Matrix{Float64})

It computes element metric

X : nodal coordinates

∂N∂a : shape functions master derivatives

∂N∂x : shape functions sparial derivatives

nn : nodes number

Jac : ∂x∂a

Jac⁻¹: ∂a∂x

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:H20}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:H20}} : twenty-nodes quadratic hexahedron identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:H27}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:H27}} : twentyseven-nodes quadratic hexahedron identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:HE8}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:HE8}} : eight-nodes linear hexahedron identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:P15}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:P15}} : fifteen-nodes quadratic prism identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:P18}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:P18}} : eighteen-nodes quadratic prism identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:PE6}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:PE6}} : six-nodes linear prism identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:T10}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:T10}} : ten-nodes quadratic tetrahedron identifier

MORFEInvariantManifold.print_cell_nodes — Method

printcelltype(io,iΩ,set,::Type{Val{:TE4}})

It writes to the output file the nodes of an element .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:TE4}} : four-nodes linear tetrahedron identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:H20}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:H20}} : twenty-nodes quadratic hexahedron identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:H27}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:H27}} : twentyseven-nodes quadratic hexahedron identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:HE8}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:HE8}} : eight-nodes lienar hexahedron identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:P15}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:P15}} : fifteen-nodes quadratic prism identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:P18}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:P18}} : fifteen-nodes quadratic prism identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:PE6}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:PE6}} : six-nodes lienar prism identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:T10}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:T10}} : ten-nodes quadratic tetrahedron identifier

MORFEInvariantManifold.print_cell_type — Method

printcelltype(io,iΩ,set,::Type{Val{:TE4}})

It writes to the output file the correct cell type assuming .vtk format.

io : file handler

iΩ : reference domain

set : reference element number

::Type{Val{:TE4}} : four-nodes linear tetrahedron identifier

MORFEInvariantManifold.read_mesh — Method

readmesh( meshfile::String, Ωlist::Vector{Vector{Int64}}, Γlist::Vector{Vector{Int64}}, mat::Vector{material}, bcdof::Vector{Vector{Int64}}, bcvals::Vector{Vector{Float64}})

It reads the mesh file and it inializes the grid data structure

mesh_file : name of the mesh file to read

Ω_list : list of domains in which the structures is organised

Γ_list : list of boundaries

mat : list of materials associated to each domain

bc_dof : list of constrained degrees of freedom of each boundary

bc_vals : boundary condition values

MORFEInvariantManifold.realification! — Method

realification!(Cp,Wr,fr,p,ndofs,neq,rmat)

It realifies mappings and reduced dynamics for an order p of the parametrisation

Cp : Parametrisation data structure

Wr : real-valued mapping

fr : real-valued reduced dynamics

p : order of the asymptotic development

ndofs : number of degrees of freedom

neq : number of equations

rmat : realification matrix R

MORFEInvariantManifold.recursive_assembly! — Function

recursive_assembly!(Cp,ndofs,p,mesh,U,ic,cc,nls=0)

It assembles all right hand sides of the homological equations for a given order p of the expansion of the ε⁰ development

Cp : parametrisation data structure

ndofs : number of degrees of freedom

p : order of the asymptotic development

mesh : mesh data structure

U : dispalcement field

ic : index combination of a given monomial

cc : counter of the asymptotic development

nls : tag for the nonlinear static analysis. 0 = yes, 1 = yes

MORFEInvariantManifold.recursive_assembly_veps! — Function

recursiveassemblyveps!(Cp,Cp⁺,ndofs,p,mesh,U,Cic,cc,nls=0)

It assembles right hand sides of all homological equations for a given order p of the expansion.

Cp : autonomous parametrisation data structure

Cp⁺ : non-auonomous parametrisation data structure

ndofs : number of degrees of freedom

p : order of the asymptotic development

mesh : grid data structure

U : displacement field

Cic : monomials index combination

cc : recursion depth

nls : nonlinear static solution. 0 = no, 1 = yes.

MORFEInvariantManifold.recursive_pm! — Method

recursivepm!(Cp,Wr,fr,Cic,Ric,pmcomb,p,ndofs,neq,comb,cc)

It realifies mappings and reduced dynamics of a given index combination comb of order p.

Cp : parametrisation data structure

Wr : real-valued mappings

fr : real-valued reduced dynamics

Cic : complex-valued parametrisation index combination of a given monomial

Ric : real-valued parametrisation index combination of a given monomial

pm_comb : +- combinations

p : order of the asymptotic development

ndofs : number of degrees of freedom

neq : number of equations

comb : reference index combination of the realified monomial

cc : recursion depth

MORFEInvariantManifold.save_matcont_rdyn — Method

savematcontrdyn(rdyn,ndofs,outdir)

It saves the autonomous reduced dynamics in a format suitable for matcont GUI usage.

rdyn : reduced dynamics in matcont format style

ndofs : number of degrees of freedom

outdir : output directory

MORFEInvariantManifold.save_matcont_rdyn_nonautonomous — Method

savematcontrdyn(rdyn,ndofs,outdir)

It saves the non-autonomous reduced dynamics in a format suitable for matcont GUI usage.

rdyn : reduced dynamics in matcont format style

ndofs : number of degrees of freedom

outdir : output directory

MORFEInvariantManifold.solve_homological! — Method

solvehomological!(Cp,ndofs,p,sysmat,sysrhs,sysres,ic,style,M,C,K)

It solves al homological equations for a given order of the asymtptotic developmet of the ε⁰-invariance.

Cp : parametrisation structure

ndofs : number of degrees of freedom

p : order of the asymptotic developmet

sys_mat : matrix used to solve the homological equations

sys_rhs : right hand side of the homological equations

sys_res : residual of the homological equations

ic : index combination

style : parametrisation style

M : mass matrix

C : damping matrix

K : stiffness matrix

MORFEInvariantManifold.solve_homological_veps! — Method

solvehomologicalveps!(η,Cp,Cp⁺,ndofs,p,sysmat,sysrhs,sys_res,style,ic,M,C,K)

It solves al homological equations for a given order of the asymtptotic developmet of the ε¹-invariance.

η : eigenvalue associated to the non-autonomous forcing

Cp : autonomous parametrisation data structure

Cp⁺ : non-auonomous parametrisation data structure

ndofs : number of degrees of freedom

p : order of the asymptotic development

sys_mat : homological equation matrix

sys_rhs : homological equation right hand side

sys_res : homological equation residual

style : parametrisation style

ic : index combination

M : mass matrix

C : damping matrix

K : stiffness matrix

MORFEInvariantManifold.structure_odir — Function

structure_odir(odir::String, id::Int64, tag = 'a')

It creates the sub-folders that stores the output of the analysis

odir : global output folder path

id : number of computed step. Autonomous is zero, non-autonomous in increasing order

tag : 'a' = autonomous, 'na' = non-autonomous

MORFEInvariantManifold.update_field! — Method

update_field!(ϕ::Field,U::Vector{Float64})

it overwrites free dofs values using the solution U

ϕ : field to update

U : solution vector

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:H20}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:H20}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:H27}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:H27}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:HE8}})

It fills ∂N∂a with the derivatives of the shape functions of a H8 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:HE8}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:P15}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:P15}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:P18}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:P18}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:PE6}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:PE6}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:T10}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:T10}} : element type specifier

MORFEInvariantManifold.∂N∂a! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:TE4}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:TE4}} : element type specifier

MORFEInvariantManifold.∂N∂a27! — Method

∂N∂a!(∂N∂a::Array{Float64},gp::Tuple,::Type{Val{:H27}})

It fills ∂N∂a with the derivatives of the shape functions of a T4 element evaluated at gauss point gp.

∂N∂a : shape functions derivatives Array

gp : gauss point coordinates

:Type{Val{:H27}} : element type specifier